+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Added `deploy --build [project name]` for quickly building project assets
+
 ## [3.7] - 05-03-2018
 ### Added
 - Digests now include information about recent malware scans, uptime, latency, and backup stats